Extract ConsistencyChecker from test-consistency
test-consistency.js exposes a function that touches the file system.
Testing with the file system is a pain; instead, I've moved
ConsistencyChecker to its own module and tested it from there.
